+++ yew-router/examples/guide/src/lib.rs
@@ -1,9 +1,0 @@
-mod guide;
-mod markdown;
-mod markdown_window;
-mod page;
-
-pub use crate::{
-    guide::{Guide, GuideProps},
-    page::{Page, PageProps},
-};