+++ yew/src/html/mod.rs
@@ -7,14 +7,13 @@ mod listener;
 mod scope;
 
 pub use listener::*;
-pub use scope::Scope;
-pub(crate) use scope::{ComponentUpdate, HiddenScope};
+pub(crate) use scope::ComponentUpdate;
+pub use scope::{AnyScope, Scope};
 
 use crate::callback::Callback;
 use crate::virtual_dom::{VChild, VList, VNode};
 use cfg_if::cfg_if;
 use cfg_match::cfg_match;
-use std::any::TypeId;
 use std::cell::RefCell;
 use std::fmt;
 use std::rc::Rc;