+++ yew/src/virtual_dom/vtag.rs
@@ -1,7 +1,7 @@
 //! This module contains the implementation of a virtual element node `VTag`.
 
 use super::{Attributes, Listener, Listeners, Patch, Reform, Transformer, VDiff, VList, VNode};
-use crate::html::NodeRef;
+use crate::html::{AnyScope, NodeRef};
 use crate::utils::document;
 use cfg_if::cfg_if;
 use cfg_match::cfg_match;
@@ -349,6 +349,7 @@ impl VDiff for VTag {
     /// to compute what to patch in the actual DOM nodes.
     fn apply(
         &mut self,
+        scope: &AnyScope,
         parent: &Element,
         previous_sibling: Option<&Node>,
         ancestor: Option<VNode>,
@@ -448,7 +449,7 @@ impl VDiff for VTag {
 
         // Process children
         self.children
-            .apply(&element, None, ancestor.map(|a| a.children.into()));
+            .apply(scope, &element, None, ancestor.map(|a| a.children.into()));
 
         let node = self.reference.as_ref().map(|e| {
             let node = cfg_match! {
@@ -838,6 +839,7 @@ mod tests {
         #[cfg(feature = "web_sys")]
         let document = web_sys::window().unwrap().document().unwrap();
 
+        let scope = AnyScope::default();
         let div_el = document.create_element("div").unwrap();
         let namespace = SVG_NAMESPACE;
         #[cfg(feature = "web_sys")]
@@ -849,17 +851,17 @@ mod tests {
         let mut svg_node = html! { <svg>{path_node}</svg> };
 
         let svg_tag = assert_vtag(&mut svg_node);
-        svg_tag.apply(&div_el, None, None);
+        svg_tag.apply(&scope, &div_el, None, None);
         assert_namespace(svg_tag, SVG_NAMESPACE);
         let path_tag = assert_vtag(svg_tag.children.get_mut(0).unwrap());
         assert_namespace(path_tag, SVG_NAMESPACE);
 
         let g_tag = assert_vtag(&mut g_node);
-        g_tag.apply(&div_el, None, None);
+        g_tag.apply(&scope, &div_el, None, None);
         assert_namespace(g_tag, HTML_NAMESPACE);
         g_tag.reference = None;
 
-        g_tag.apply(&svg_el, None, None);
+        g_tag.apply(&scope, &svg_el, None, None);
         assert_namespace(g_tag, SVG_NAMESPACE);
     }
 
@@ -980,6 +982,7 @@ mod tests {
 
     #[test]
     fn it_does_not_set_empty_class_name() {
+        let scope = AnyScope::default();
         let parent = document().create_element("div").unwrap();
 
         #[cfg(feature = "std_web")]
@@ -988,7 +991,7 @@ mod tests {
         document().body().unwrap().append_child(&parent).unwrap();
 
         let mut elem = html! { <div class=""></div> };
-        elem.apply(&parent, None, None);
+        elem.apply(&scope, &parent, None, None);
         let vtag = assert_vtag(&mut elem);
         // test if the className has not been set
         assert!(!vtag.reference.as_ref().unwrap().has_attribute("class"));
@@ -996,6 +999,7 @@ mod tests {
 
     #[test]
     fn it_does_not_set_missing_class_name() {
+        let scope = AnyScope::default();
         let parent = document().create_element("div").unwrap();
 
         #[cfg(feature = "std_web")]
@@ -1004,7 +1008,7 @@ mod tests {
         document().body().unwrap().append_child(&parent).unwrap();
 
         let mut elem = html! { <div></div> };
-        elem.apply(&parent, None, None);
+        elem.apply(&scope, &parent, None, None);
         let vtag = assert_vtag(&mut elem);
         // test if the className has not been set
         assert!(!vtag.reference.as_ref().unwrap().has_attribute("class"));
@@ -1012,6 +1016,7 @@ mod tests {
 
     #[test]
     fn it_sets_class_name() {
+        let scope = AnyScope::default();
         let parent = document().create_element("div").unwrap();
 
         #[cfg(feature = "std_web")]
@@ -1020,7 +1025,7 @@ mod tests {
         document().body().unwrap().append_child(&parent).unwrap();
 
         let mut elem = html! { <div class="ferris the crab"></div> };
-        elem.apply(&parent, None, None);
+        elem.apply(&scope, &parent, None, None);
         let vtag = assert_vtag(&mut elem);
         // test if the className has been set
         assert!(vtag.reference.as_ref().unwrap().has_attribute("class"));
@@ -1067,6 +1072,7 @@ mod tests {
 
     #[test]
     fn swap_order_of_classes() {
+        let scope = AnyScope::default();
         let parent = document().create_element("div").unwrap();
 
         #[cfg(feature = "std_web")]
@@ -1075,7 +1081,7 @@ mod tests {
         document().body().unwrap().append_child(&parent).unwrap();
 
         let mut elem = html! { <div class=("class-1", "class-2", "class-3")></div> };
-        elem.apply(&parent, None, None);
+        elem.apply(&scope, &parent, None, None);
 
         let vtag = if let VNode::VTag(vtag) = elem {
             vtag
@@ -1101,7 +1107,7 @@ mod tests {
         } else {
             panic!("should be vtag")
         };
-        vtag.apply(&parent, None, Some(VNode::VTag(ancestor)));
+        vtag.apply(&scope, &parent, None, Some(VNode::VTag(ancestor)));
 
         let expected = "class-3 class-2 class-1";
         assert_eq!(get_class_str(&vtag), expected);
@@ -1117,6 +1123,7 @@ mod tests {
 
     #[test]
     fn add_class_to_the_middle() {
+        let scope = AnyScope::default();
         let parent = document().create_element("div").unwrap();
 
         #[cfg(feature = "std_web")]
@@ -1125,7 +1132,7 @@ mod tests {
         document().body().unwrap().append_child(&parent).unwrap();
 
         let mut elem = html! { <div class=("class-1", "class-3")></div> };
-        elem.apply(&parent, None, None);
+        elem.apply(&scope, &parent, None, None);
 
         let vtag = if let VNode::VTag(vtag) = elem {
             vtag
@@ -1151,7 +1158,7 @@ mod tests {
         } else {
             panic!("should be vtag")
         };
-        vtag.apply(&parent, None, Some(VNode::VTag(ancestor)));
+        vtag.apply(&scope, &parent, None, Some(VNode::VTag(ancestor)));
 
         let expected = "class-1 class-2 class-3";
         assert_eq!(get_class_str(&vtag), expected);